+++ examples/basic/hosts/hm-standalone.x86_64-linux/default.nix
@@ -1,5 +1,0 @@
-_: {
-  boot.isContainer = true;
-  networking.hostName = "hm-standalone";
-  system.stateVersion = "25.05";
-}