+++ examples/basic/hosts/hm-standalone.x86_64-linux/meta.nix
@@ -1,4 +1,0 @@
-{
-  roles = [ "server" ];
-  home.embed = false;
-}