+++ examples/basic/hosts/nixos-desktop.x86_64-linux/meta.nix
@@ -1,8 +1,0 @@
-{
-  roles = [
-    "desktop"
-    "development"
-  ];
-
-  home.useGlobalPkgs = false;
-}